requirements.txt
selene>=2.0.0rc9
pytest
python-dotenv
pydantic
pydantic-settings
black
appium-python-client
allure-pytest
requests